+++ src/k8s-extension/azext_k8s_extension/partner_extensions/OpenServiceMesh.py
@@ -4,21 +4,37 @@
 # --------------------------------------------------------------------------------------------
 
 # pylint: disable=unused-argument
+# pylint: disable=redefined-outer-name
+# pylint: disable=no-member
 
-from azure.cli.core.azclierror import InvalidArgumentValueError, RequiredArgumentMissingError
 from knack.log import get_logger
 
-from ..vendored_sdks.models import ExtensionInstance
-from ..vendored_sdks.models import ExtensionInstanceUpdate
-from ..vendored_sdks.models import ScopeCluster
-from ..vendored_sdks.models import Scope
+from azure.cli.core.azclierror import InvalidArgumentValueError, RequiredArgumentMissingError
+from azure.cli.core.commands.client_factory import get_subscription_id
+
+from packaging import version
+import yaml
+
+from ..partner_extensions import PartnerExtensionModel
 
 from .PartnerExtensionModel import PartnerExtensionModel
+
+from ..vendored_sdks.models import (
+    ExtensionInstance,
+    ExtensionInstanceUpdate,
+    ScopeCluster,
+    Scope
+)
+
+from .._client_factory import cf_resources
 
 logger = get_logger(__name__)
 
 
 class OpenServiceMesh(PartnerExtensionModel):
+    CHART_NAME = "osm-arc"
+    CHART_LOCATION = "https://azure.github.io/osm-azure"
+
     def Create(self, cmd, client, resource_group_name, cluster_name, name, cluster_type, extension_type,
                scope, auto_upgrade_minor_version, release_train, version, target_namespace,
                release_namespace, configuration_settings, configuration_protected_settings,
@@ -38,30 +54,21 @@ class OpenServiceMesh(PartnerExtensionModel):
         scope_cluster = ScopeCluster(release_namespace=release_namespace)
         ext_scope = Scope(cluster=scope_cluster, namespace=None)
 
-        valid_release_trains = ['staging', 'pilot']
-        # If release-train is not input, set it to 'stable'
-        if release_train is None:
+        # version is a mandatory if release-train is staging or pilot
+        if version is None:
             raise RequiredArgumentMissingError(
-                "A release-train must be provided.  Valid values are 'staging', 'pilot'."
+                "A version must be provided for release-train {}.".format(release_train)
             )
-
-        if release_train.lower() in valid_release_trains:
-            # version is a mandatory if release-train is staging or pilot
-            if version is None:
-                raise RequiredArgumentMissingError(
-                    "A version must be provided for release-train {}.".format(release_train)
-                )
-            # If the release-train is 'staging' or 'pilot' then auto-upgrade-minor-version MUST be set to False
-            if auto_upgrade_minor_version or auto_upgrade_minor_version is None:
-                auto_upgrade_minor_version = False
-                logger.warning("Setting auto-upgrade-minor-version to False since release-train is '%s'", release_train)
-        else:
-            raise InvalidArgumentValueError(
-                "Invalid release-train '{}'.  Valid values are 'staging', 'pilot'.".format(release_train)
-            )
+        # If the release-train is 'staging' or 'pilot' then auto-upgrade-minor-version MUST be set to False
+        if auto_upgrade_minor_version or auto_upgrade_minor_version is None:
+            auto_upgrade_minor_version = False
+            logger.warning("Setting auto-upgrade-minor-version to False since release-train is '%s'", release_train)
 
         # NOTE-2: Return a valid ExtensionInstance object, Instance name and flag for Identity
         create_identity = False
+
+        # _validate_tested_distro(cmd, resource_group_name, cluster_name, version)
+
         extension_instance = ExtensionInstance(
             extension_type=extension_type,
             auto_upgrade_minor_version=auto_upgrade_minor_version,
@@ -93,3 +100,63 @@ class OpenServiceMesh(PartnerExtensionModel):
             release_train=release_train,
             version=version
         )
+
+    def Delete(self, client, resource_group_name, cluster_name, name, cluster_type):
+        pass
+
+
+# def _validate_tested_distro(cmd, cluster_resource_group_name, cluster_name, extension_version):
+
+#     field_unavailable_error = '\"testedDistros\" field unavailable for version {0} of microsoft.openservicemesh, ' \
+#         'cannot determine if this Kubernetes distribution has been properly tested'.format(extension_version)
+
+#     if version.parse(str(extension_version)) <= version.parse("0.8.3"):
+#         logger.warning(field_unavailable_error)
+#         return
+
+#     subscription_id = get_subscription_id(cmd.cli_ctx)
+#     resources = cf_resources(cmd.cli_ctx, subscription_id)
+
+#     cluster_resource_id = '/subscriptions/{0}/resourceGroups/{1}/providers/Microsoft.Kubernetes' \
+#         '/connectedClusters/{2}'.format(subscription_id, cluster_resource_group_name, cluster_name)
+
+#     resource = resources.get_by_id(cluster_resource_id, '2020-01-01-preview')
+#     cluster_distro = resource.properties['distribution'].lower()
+
+#     if cluster_distro == "general":
+#         logger.warning('Unable to determine if distro has been tested for microsoft.openservicemesh, '
+#                        'kubernetes distro: \"general\"')
+#         return
+
+#     tested_distros = _get_tested_distros(extension_version)
+
+#     if tested_distros is None:
+#         logger.warning(field_unavailable_error)
+#     elif cluster_distro not in tested_distros.split():
+#         logger.warning('Untested kubernetes distro for microsoft.openservicemesh, Kubernetes distro is %s',
+#                        cluster_distro)
+
+
+# def _get_tested_distros(chart_version):
+
+#     try:
+#         chart_arc = ChartBuilder({
+#             "name": OpenServiceMesh.CHART_NAME,
+#             "version": str(chart_version),
+#             "source": {
+#                 "type": "repo",
+#                 "location": OpenServiceMesh.CHART_LOCATION
+#             }
+#         })
+#     except VersionError:
+#         raise InvalidArgumentValueError(
+#             "Invalid version '{}' for microsoft.openservicemesh".format(chart_version)
+#         )
+
+#     values = chart_arc.get_values()
+#     values_yaml = yaml.load(values.raw, Loader=yaml.FullLoader)
+
+#     try:
+#         return values_yaml['OpenServiceMesh']['testedDistros']
+#     except KeyError:
+#         return None